pyproject.toml
[tool.poetry]
name = "quotes-dab"
version = "0.1.0"
description = ""
authors = ["doug-pires <dpiresmartins@yahoo.com.br>"]
readme = "README.md"
packages = [{include = "quotes_dab"}]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
databricks-sdk = "^0.11.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-mock = "^3.11.1"
black = "^23.9.1"
isort = "^5.12.0"
chispa = "^0.9.4"
pytest-cov = "^4.1.0"
pyspark = "3.4"
delta-spark = "2.4"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.4.6"
mkdocstrings = "^0.23.0"
mkdocstrings-python = "^1.7.3"
pymdown-extensions = "^10.3.1"
mkdocs-macros-plugin = "^1.0.4"


[tool.pytest.ini_options]
addopts = ["--strict-markers", "--strict-config", "-ra","-v","--setup-show","-s"]
testpaths = ["tests"]

[tool.coverage.report]
exclude_lines = [
    "if self.debug:",
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]

[tool.coverage.run]
omit = [
    "quotes/create_catalog_quotes.py"
]

[tool.black]
line-length = 88

[tool.poetry.scripts]
request-quote = "quotes_dba.request_quote:main"
main = "quotes_dba.dummy:main"